Move detect_deadlocks to main.rs
